Refactor active_sensor_pixel_dimensions
